Camera : HAL : Return proper information about the camera hardware to upper layers
This changes the HAL behaviour when a camera is not started. It used
to still send a reference to the camera object even though the camera hardware
was not present. Now it returns NULL if camera is not started and passes the
reference to camera object only when the camera is successfully started.