platform/hardware/msm7k: Do not close snd driver for 7k until phone powerdown.
Opening and Closing sound driver often (for each call made to snd driver) results in wrong device routing. It also results in the failure of mute / un-mute of voice calls. The sound driver is created at the phone bootup and the same instance of the driver is used until the phone powerdowns, which fixes these issues.

CRs-fixed: 183111, 182586