Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I'd still suggest cleaning up the whitespace, but the code is working for me now (at least on Inc, I don't have others to test).  As described on XDA[1], a few of us had an issue that you must first select a custom color before even the regular colors will work, but then it works fine.  That may not be related to this patch per se... it's just bizarre, something to note.

[1] http://forum.xda-developers.com/showthread.php?t=847317

Patch-set: 2