Update patch set 2
Patch Set 2:

We could do that, but many phones have the amber/green led.. what's the point of reimplementing it for each device?

On another note, which phones use msm7k, but do have rgb? Maybe we can make separate liblights and put build strings to specificy which one to build. It seems like supersonic tried to do this but never finished implementing it.

I'm hoping someone else will commit to this fix.

Patch-set: 2