Update patch set 1
Patch Set 1: Verified; Looks good to me, but someone else must approve

I went ahead and verified it because everyone seems to agree that it fixes it (I'm unable to test).  I think we should make a 7.0.4 including this fix. (and still 2.3.3)  I'll push it through gingerbread-release once it gets approved.

Patch-set: 1