Update patch set 1
Patch Set 1:

This was reverted a few months ago in commit:

https://github.com/CyanogenMod/android_hardware_msm7k/commit/a2be4dede6a0ed259441b92c9f4b896b4ca146e4

I believe the reason for it being reverted, if I remember correctly, was that it broke output to speaker for FM radio.
This should now be fixed with http://review.cyanogenmod.com/#change,8331.

I noticed the following in the log though after this patch, but it doesn't seem to cause any harm. Feel free to look into it though if anyone wants.

W/AudioPolicyManager(  125): getDeviceForStrategy() unknown strategy: 4

Patch-set: 1