Update patch set 2
Patch Set 2:

Kali, I know we've discussed this at length on IRC, but perhaps we should take this change for now until a full solution can be created. On Saga this change completely removes the sound hang, and you yourself said it reduces it in your tests.

I get many people asking me why VoIP kills the sound and requires a reset. Surely a temporary fix is better than leaving it in the current broken state?

Patch-set: 2