Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I advise against BOARD_USES_QCOM_HARDWARE, because CAF uses this flag to sneak in their proprietary libs, which we don't have. It would just be a precaution in case we pull one of their branches wholesale such as Webkit or Chromium.

Patch-set: 1