Update patch set 1
Patch Set 1:

I fully support the change, but could we move it into a "qcom" directory for better categorizing of qcom-specific components? In my local source it is lumped with hardware/qcom/media next to libstagefrighthw. It isn't necessarily "media", but it places all qcom libs together out of msm7k.

Patch-set: 1