Update patch set 2
Patch Set 2:

I had made this statement previously and to save the "go look at CAF" explanation it is because the version in msm7k contains the audio configuration, which is actually more updated and ready for the audio than the version currently in media. The header for MetaData, on the other hand is directly referenced in this location throughout the frameworks updates. It seems stupid, but it is actually the easiest way about it. Hopefully that helps to explain it better.

Patch-set: 2