Update patch set 1
Patch Set 1:

OK, yeah this is a hack, as I mentioned in the bug report, which just happens to make one particular use case work for me. It was suggested there to submit it here.

If you'll be able to take a look at it that would be great, otherwise I can dig deeper to gain a better understanding of this class and how it's used. If you don't have access to HTC Evo Shift 4G, I'll be happy to help with testing.

Patch-set: 1